Move load_model_params to utils module
